augmented_manifest: store null p1/p2 as - instead of empty string
Summary:
To make it easier to understand and parse the p1 and p2 hashes, encode the null values as - (in the same way as we do for the computed node id when it matches the node id).

This is a format change, but we haven't derived any of these yet, so this should be ok.

Reviewed By: liubov-dmitrieva

Differential Revision: D58180872

fbshipit-source-id: ee11962dd08985e6d285be7f59c75bb2ff8882fd