blake3: change key constant to a byte array
Summary: The `blake3_constants::BLAKE3_HASH_KEY` constant is defined as a string slice and is then converted to a 32-byte byte array at runtime.  This results in unnecessary copying and bounds checking.  We can switch to a byte array constant and the type checker will enforce that it's the right size.

Reviewed By: zertosh

Differential Revision: D58137059

fbshipit-source-id: 0fa09b86a1c35cf7ce100e021a841744ca057f05