Update integration tests to better reproduce repo merge operation
Summary:
Currently the helpers that set up submodule expansion repos for integration tests use a single commit sync version in the entire process.

We have the submodule repo with revisions A-B-....-F-G, where `G` is the latest commit in the submodule repo before the import and merge into large repo.

## What happens now
We run the `initial-import` command passing commit `G`. So it imports all the commits using the noop config version.
Then, we create the merge commits using the synced version of `G`, i.e.`G'`.

## What happens in production
We run the initial import command for commit `G`, getting `G'` in large repo.

We wait until a new commit, let's call it `H` is committed to submodule repo.
This commit `H` is then used with the `monononke_x_repo_sync_job once` command along with the `--unsafe-change-version-to` flag to change the mapping version. So commit `H'` is what changes the config version for the entire large repo.

## What happens after this change
We run the initial import until commit `F` and leave the commit `G` so that it plays the role of commit `H` in production, i.e. we sync `G` after the merge was done to change the mapping.

Reviewed By: mitrandir77

Differential Revision: D58182645

fbshipit-source-id: d0ea5088206931746ec4395288fefa66cb095d7d