tests: add a Buck target for testing sl + eden + watchman
Summary:
There are some cases when we want to test the three mentioned in the title at the same time.
Unlike the EdenFS or Watchman Buck targets, however, we don't want all tests to be tested
with Watchman + EdenFS since that might be slow to run. These will be opt in and should be
added throug the `get_hg_edenfs_watchman_run_tests_included` function.

Reviewed By: zzl0

Differential Revision: D58160646

fbshipit-source-id: feb437d2537dabb01aca4d1b8724ab80f66997dc