manifest: add comparison method for more generalized comparisons of sharded manifest
Summary:
Add a new way to compare manifests with their parents, that is optimized for sharded manifests.

With these comparisons, there are now 7 possible variants:

  * `New`: A single new entry.
  * `Same`: A single entry that is the same as one of the parents.
  * `Changed`: A single entry that has changed relative to all of the parents.
  * `Removed`: A single entry that was present in at least one parent but is now removed.
  * `ManyNew`: A group of new entries with a common prefix.
  * `ManySame`: A group of entries with a common prefix that are all the same as one of the parents.
  * `ManyRemoved`: A group of entries with a common prefix that were present in at least one parent but are now removed.

The `Many*` variants can operate on large subsections of sharded manifests, allowing callers to efficiently process them.  For example, with `ManySame`, the value from the parent can be re-used without needing to load it all.

Reviewed By: liubov-dmitrieva

Differential Revision: D58154702

fbshipit-source-id: 06cad5efb12376c8e879f294b1a9303de8d2a893