debugbisect: command dedicated for stable bisect testing
Summary:
The bisect tests is tricky because the `bisect -g` or `bisect -b` heavily
depends on context. For example, a test case might look like:

  $ hg bisect -r
  $ hg bisect -g 33b1f9bc8bc58c05855524ce9c1a69d916ac05f2
  $ hg bisect -b 'desc(17)'   # -> update to rev 6
  $ hg bisect -g      # -> update to rev 13
  Testing changeset b0a32c86eb31 (9 changesets remaining, ~3 tests)
  3 files updated, 0 files merged, 1 files removed, 0 files unresolved
  $ hg bisect -s      # -> update to rev 10
  Testing changeset 429fcd26f52d (9 changesets remaining, ~3 tests)
  3 files updated, 0 files merged, 1 files removed, 0 files unresolved
  $ hg bisect -b      # -> update to rev 8
  Testing changeset dab8161ac8fc (3 changesets remaining, ~1 tests)
  2 files updated, 0 files merged, 1 files removed, 0 files unresolved
  $ hg bisect -g      # -> update to rev 9
  Testing changeset 3c77083deb4a (2 changesets remaining, ~1 tests)
  1 files updated, 0 files merged, 0 files removed, 0 files unresolved
  $ hg bisect -b
  The first bad revision is:
  commit:      3c77083deb4a
  user:        test
  date:        Thu Jan 01 00:00:09 1970 +0000
9

Any slight change in the bisect strategy (which is complex when there are
merges) like picking a different commit, or having less/more bisect steps would
break the test.

This diff adds a dedicated command to output all bisect steps in one go, so if
the bisect algorithm changes, the test update can be just a single `-i` run
without manually editing lots of bisect commands.

Reviewed By: muirdm

Differential Revision: D58045490

fbshipit-source-id: fb790bf82401ec21ba6856d8db40c6f654b82106