ui: make popbuffer accept bytes
Summary:
Somehow I got the `ui.pushbuffer` mixed with str and bytes in the next diff:

  [..., 'The first bad revision is:\n', b'A16\n']

It does not seem necessary to enforce the `str` type on the buffer, given that
`ui.write` and `ui.writebytes` can be used (i.e. calling `ui.write` does not
make follow-up `ui.writebytes` error out). So let's relax the `popbuffer` and
`popbufferbytes` to convert between str and bytes automatically.

Reviewed By: muirdm

Differential Revision: D58045491

fbshipit-source-id: 8ddf62cf04d7d95e949228b8f25a9f4b9621d860