eagerepo: don't return "history" error for unknown files
Summary:
Tweak eagerepo behavior to match Mononoke by not propagating 404 errors when requesting file history information for an unknown file node.

I'm trying to make "rebase" re-use the filelog entry of unmerged (i.e. unchanged) files in the rebased commits. This was causing an eagerepo error due to this sequence of events:
1. Make local commit A, don't push to server.
2. Rebase A and attempt to reuse filenodes.
3. Dirsync happens, copying A/foo to A/bar (using the same filenode).
4. repo.commitctx() calls metadatastore.prefetch(keys), which doesn't find `(A/bar, <node>)` locally (metadata store is keyed by file path and node), so it makes a remote request.
5. Eagerepo doesn't know about the file since the commit wasn't pushed.

Reviewed By: quark-zju

Differential Revision: D58022949

fbshipit-source-id: 45dc927e87f3f3d84c741c03b8591fc4711c21bd