context: make ctx.write accept "bytes | filectx"
Summary:
This allows for the context to make more optimizations, depending on what it is doing.

In particular, I want to make the in-memory rebase reuse the existing file node if the file isn't changing, so having access to the file context object (complete with file node) will be useful.

Reviewed By: zzl0

Differential Revision: D57972716

fbshipit-source-id: e383593bef132e3b52ce7a66a3946dfd06fb7913