context: remember file node in overlayworkingctx
Summary:
overlayworkingctx is used by in-memory rebase to intercept "write()" operations. In a previous diff, I made the "write()" operations receive the filectx instead of bytes, when available. In this diff, we notice that we got a filectx and store the file node for later use.

I changed the overlay cache to store the filectx_or_bytes as "data" instead of bytes so the "fctx.data()" call becomes lazy.

Reviewed By: quark-zju

Differential Revision: D57972713

fbshipit-source-id: ed04ead27d9892eb10827497e5409f74d04041fb