Become resilient to interrupted derivation
Summary:
This is currently one of the worst UX pain-points of gitimport:
If the previous import (from gitimport or from Live Git Sync) got interrupted uncleanly for some reason,
data derivation will fail due to the rigid requirements of `derive_bulk` that the parents must already
have been derived.

It happened to me again yesterday on `chromium/src` after the Live Git Sync rejected a push and the
repo was in an inconsistent state.

NOTE: Work is being done towards making the data derivation interface handle this case, but
until that is done, we're just implementing it locally.

Reviewed By: RajivTS

Differential Revision: D58284780

fbshipit-source-id: c15c2338c91c806ef7da492318fe5b7aa0443953