simplify PACKAGE files
Summary:
## Rationale

This diff switches our Citadel CI setup from opt-in `dev-rust-oss` build to opt-out
one. I'm turning that build mode for entire mononoke directory and removing all
custom autogenerated opt-ins we had in our subdirectories. There are 2 good
reasons to make this change:

1. We need 7 new PACKAGE files with opt-outs in place of 184 files with  opt-ins.
2. If we forget to opt-out something it will be visible as build failre and red
signal on diffs.  If we forget to opt-in something we won't be notified (and we didn't notice in the past).
Explicit signal is always better than lack of it.

My ulterior motive is to prevent oss builds from not running. Right now for example all integration tests were not running in oss mode.

## How

Removed all the PACKAGE files that only add dev-rust-oss mode to the ci build
using this command:

```
$ cat cmdlib/environment/PACKAGE | sha1sum
4f6b36727f455de60cd28615ed015bdc6fa40bbe  -
$  cat cmdlib/environment/PACKAGE
load("fbcode//target_determinator/macros:ci.bzl", "ci")

ci.package(
    ci.linux(
        ci.mode("fbcode//mode/dev-rust-oss"),
    ),
)
$ fd PACKAGE | xargs -I {} -- bash -c '[[ "$(sha1sum {}  | cut -f 1 -d " ")" == "4f6b36727f455de60cd28615ed015bdc6fa40bbe" ]] && rm {}'
```

instead added that mode into top-level file and added opt-outs in all `facebook/` directories that contained buck targets

This plan is not fully bulletproof: there might have been something not
opted-in for a reason but we'll learn about that soon enough.

Reviewed By: singhsrb

Differential Revision: D58287931

fbshipit-source-id: 5b0f8508cbdd78602fd931cef45594abb5081610