[reset split from suggestion after send
Summary:
# Context
I realized that nothing is clearing the split suggestion tracking in the split dialog and also that if you split the suggestion, then cancel and go to the other it would retain the old value.

basically the tracking was totally fugashed, so this fixes it.

I thought about changing the logging to be an entry point string value, I could be convinced to do that but this seems simple enough for my use case.

Reviewed By: evangrayk

Differential Revision: D58264521

fbshipit-source-id: d3e20c159459d2a88862326ddbb5794648c949b0