Add globFiles to SaplingNativeBackingStore
Summary:
This diff starts bringing over globFiles into the C++ side of the backing store.

std::array<unsigned char, 20> is the type generated by the ffi for the commit id.
The input type from GlobParams is a ThriftRootId(thrift type, c++ generated type string), so will need to be converted. Accepting opinions on whether it should be done in this layer or a previous one.

Reviewed By: kmancini

Differential Revision: D56958618

fbshipit-source-id: 62ac2e7d3f462d0d5d5b3b740d258e3a14331071