rebase: re-use filelog entries for on-disk commits
Summary:
In the previous commit I made rebase in-memory commits re-use the filelog entry for unchanged files. In this commit I make it work for non-in-memory (i.e. on-disk) commits.

This way the "file too big" check won't fire again for files that were already committed.

I put in a config knob so we can disable this behavior if necessary. It seems safe and I would expect it to speed things up overall since rebase is so common, but it's possible it will slow things down or have some other issue.

Reviewed By: quark-zju

Differential Revision: D57972714

fbshipit-source-id: c92654194676bb5650c67b59c8f998e989806e7c