rebase: re-use filelog entries for in-memory commits
Summary: When performing an in-memory rebase, we will now re-use filelog objects for the successor commits (if the file contents haven't changed and the file parents haven't changed). This might speed things up a bit, but my actual goal was to make the file size limit error not fire again when rebasing a commit with a big file.

Reviewed By: quark-zju

Differential Revision: D57972712

fbshipit-source-id: 8cde9b1ecec898bd27104aadd787f13518518cf5