rebase: add '--quit' flag to rebase command
Summary:
Before this change, users can only either `continue` the rebase operation or `abort` the entire rebase command when there are conflicts during rebasing a large diff stack. This diff adds `rebase –quit` to quit the rebase state and keep the already rebased commits, and users can do the rest of rebase later with `sl restack` or `sl rebase -s <next_commit_in_diff_stack> -d <where_we_left>`.

This is also similar to `git rebase --quit`, which aborts the rebase operation but HEAD is not reset back to the original branch. The index and working tree are also left unchanged as a result.

Reviewed By: quark-zju

Differential Revision: D58370354

fbshipit-source-id: 596bcac6ee71c02a38a7306a798af007819886e1