scmstore: use shared error type to expose "key"
Summary:
I want to access the Key associated with errors, but currently scmstore includes the Key in its own error type (and I don't want to depend on the revisionstore crate from the checkout crate). Instead, define a new "KeyedError" shared error type, and use that in scmstore.

I simplified the error tracking to only remember the first error for a given key rather than a list of errors. I don't think we get multiple errors fetching a single key.

Reviewed By: sggutier

Differential Revision: D58216861

fbshipit-source-id: f4b191b6b4559cf84fac16875650b7c119e3d418