Enable git client to drive negotiation in Git protocol
Summary:
With the server side implementation of Mononoke commit graph, within just a single request we can figure out exactly the set of commits that needs to be sent to the client. However, Git being Git, the client doesn't advertise all the HAVEs in one go and likes to play a couple rounds of negotiation before arriving at the smallest set of commits to be sent across.

So far the server would look at the first set of HAVEs sent by the client and immediately start sending the packfile. As a result, we would sent an accurate yet unnecessarily big packfile. This diff ensures that we engage in as many rounds of negotiations as needed by the client and only start sending the packfile when the client sends the done flag.

NOTE: This problem is peculiar since it doesn't occur in every repo. The repo needs to be big enough, have large number of commits and a healthy commit rate for this issue to show up (e.g. oculus/aosp/vendor/oculus or chromium/src)

Reviewed By: gustavoavena

Differential Revision: D58291027

fbshipit-source-id: 81d41becb6f42cb9c2ed2f8ad15db34e2a7d7141