use previous for loading state in fetch pending sloc
Summary:
this diff updates the fetching of pending sloc to use the previous value to avoid the UI jumping around as the user selects / deselects files.  This also makes the UI less loading-spinner-ey.  to accomplish this i converted the pending sloc atom to use a loadable which gives the flexibility to capture loading state and track/return the previous value for that as needed.  This removes the need for a suspense boundary in the pending diff stats component now.

i was not able to use the usePrevious hook as it had to be done as part of this loading state and the typings related to how that works.

Reviewed By: evangrayk

Differential Revision: D58418890

fbshipit-source-id: 8331aedd4bcceeed0da2cc2a61f05373b46efb40