dag: add Set::specialized_flatten_id
Summary:
Right now most Set types are IdStaticSet for set operation (union,
intersection, difference) fast paths that return IdStaticSet. The problem of
that is that the Union operation no longer preserves order (which is a feature
existed in Python, but not in the Rust impl).

In the future we want the Rust impl to use UnionSet to preserve set iteration
order while maintain fast paths. This diff adds a method to obtain IdStaticSet
recursively so we might maintain fast paths using this API.

Reviewed By: zzl0

Differential Revision: D58304629

fbshipit-source-id: a3ec4a234baea535b39c4bab75b69dbc70d3e3c0