pydag: make nameset's `+` operation preserve order
Summary:
Make it possible to preserve set order like in Python. This can be potentially
break things (either behavior-wise or performance-wise). A config is added to
restore the old behavior in case it breaks things.

The `|` operation is kept unchanged - it does not preserve order.

| layer | `+` | `｜` | note |
| revsetlang | preserve order | preserve order | public API, cannot change |
| Python smartset | preserve order [1] | not defined [4] | private API, widely used |
| Binding nameset | not preserve order [2] -> preserve order | not preserve order | `repo.dageval`, private API, less widely used |
| Rust NameSet | not defined -> preserve order [3] | not preserve order | private API, very limited use |

[1]: Except for the `nameset` or `idset` backed by bindings `pydag` [2].
[2]: This was an implementation limitation, not intentional. See D22519584 for related subtle test changes.
[3]: added recently in D58308085.
[4]: plan to define it given "unordered union" seems a frequent operation.

So the near term final state is that `+` always preserves order in all layers. `|` does not preserve in all layers except for the revsetlang.
Ideally revsetlang can be updated about the `|`. However that might be potentially breaking and we need to be very careful rolling it out.

Reviewed By: zzl0

Differential Revision: D58312755

fbshipit-source-id: 3d485cfbcdada543b87d7174e4f6e823713c1819