commit cloud: better errror handling
Summary:
If something goes wrong with mononoke, the server simply  doesn't return anything, which makes it hard for me to debug any arising issues, so making this more descriptive.

Before, we returned a JSON with the requested info on success, but nothing on failure. Now we always return something that includes either the data, or an error message.  See the changes to `test-edenapi-modern-commitcloud.t` to see the wrapping added in this diff. I basically wrapped `ReferencesData` (the json data) in `ReferencesDataResponse` (a result with either the json data or an error message).  And piped that all over the EdenAPI flow.

I also modified the client (sapling) to handle both this new wrapping and the old one, cause I don't know how in sync the mononoke and the sapling release are so better to be safe.

Reviewed By: markbt

Differential Revision: D58415855

fbshipit-source-id: dceab91cc557bad35f662ba04b8e695c576c8ddc