Commit Graph: implement MemWrites storage backend
Summary:
In cross_repo_sync, we need a version of the `CommitGraph` that reads from the repo's `CommitGraph` but is guaranteed never to write to it.

It was implemented with an `unimplemented!()` version of each of the `CommitGraphStorage` trait methods. That forces the code that uses it to never use `CommitGraph`.

In D58202861, we need to actually use the `CommitGraph` to leverage its optimizations, so in D58239336, we replace the dummy implementation of this trait with this one.

But first, in this diff, we create it :)

Reviewed By: gustavoavena

Differential Revision: D58239337

fbshipit-source-id: 58ed950e5f3d6c10bd7bdd9d2ada4b1984e25f2f