bookmarks_movement: Stop doing unbounded work on bookmark creation
Summary:
Fixes a bottleneck in Live Git Sync and gitimport for chromium/src.

Before this change, the same restrictions enforcement would run again and again on public commits.
With this change, we restrict the set of commits on which to check restrictions to the subset of draft commits being pushed.

One example where this helps is this:
We have a long branch that diverged from master a long time ago. We are creating a tag on that branch.
Without this change, we would do `O(n_commits_in_long_branch)` checks per tag creation, including re-checking the same commits we already checked over and over again.
For gitimport, this becomes `O(n^2)`.
After this change, we would do 0 checks in that scenario, or `O(n_new_commits_being_pushed)` in a slightly different scenario, which is the exact set we are interested in running checks for.

Use a jk for the rollout because this code is on the hot path for lands for fbsource, so we will want to rollout the jk while monitoring push throughput to make sure we don't unintentionally cause a SEV if the delay added for finding the precise boundary is significant in this context.

Reviewed By: gustavoavena

Differential Revision: D58245003

fbshipit-source-id: 4ed8d2d542eaa92527d8667222bfee504d629586