Fix missing tag import in gitimport
Summary: As part of performance optimizations, we deferred the upload of tags to only those scenarios where the commit that they point to changes. However, there can be (and are) cases where the actual content of the tag changes without changing its name or the commit that it points to. In these cases, the current version of gitimport simply ignores the tag and we end up in an out-of-sync situation. This diff remedies that.

Differential Revision: D58468665

fbshipit-source-id: d3fb0ce9fc6e60c11586c987e94e5300a8b619e2