Stop considering untitled files as unsaved
Summary:
We would consider newly created files as unsaved and warn you about them when committing in ISL.

The reason we had this initially was to prevent people from forgetting to save their files before committing/amending.

However, if your file does not have a name at all, it seems unlikely that you already intended for it to be committed. I'm sure there are cases, but I think files which already have a file name will be much more likely.

It's quite easy for this to be a false positive since the "is part of your repo" test will often fail for untitled files. People who have scratch files will get warnings, even though there's no reason to suggest that untitled file is in your repo.

Reviewed By: quark-zju

Differential Revision: D58475992

fbshipit-source-id: 4091c1edd3ed7dbf7e7da55175d19b52b44d68cf