revisionstore: split concepts of local/shared from permanent/rotated
Summary:
Currently indexedlog stores can either be "local", which means they are a single permanent indexedlog located in the repo's local storage, or "shared", which means they are a rotated indexedlog located in the shared cache storage.

We would like to add a new data store that is stored in the local store, but is rotated.

To allow this to make sense, we need to split the two concepts.

The store *type* will be "permanent" or "rotated" indicating whether it's a single log that is never deleted from, or a rotated log where the oldest log is periodically deleted.

The store *location* will be whether the store is located in the local repository or in the shared cache.

Reviewed By: quark-zju

Differential Revision: D58435576

fbshipit-source-id: d350e6a40c17e5aaba78b5eed646a2f688125b3c