add DirectoryMetadata to TreeEntry
Summary:
# Context

The new AugmentedManifest format used by Sapling and Mononoke will support looking up Trees/Blobs from CAS via (Blake3,size) pairs. We already provide a way for users to get (Blake3,size) pairs for blobs, but there is currently no easy way to get this information for trees.

We want to add support to getAttributesFromFilesV2 for looking up the (Blake3,size) pairs for trees as well. To do this, we'll need to add a method for querying TreeMetadata (including size and Blake3 hashes) via the SaplingBackingStore.

This initial implementation of TreeMetadata will only include the following features:

1) Ability to query TreeMetadata from the SaplingBackingStore via a getTreeMetadata endpoint
2) The ability to request TreeMetadata from the getAttributesFromFilesV2 thrift endpoint

In the future, we will extend the implementation to support:

1) Automatically fetching TreeMetadata during BackingStore::getTree requests
2) Caching TreeMetadata in Eden's in-memory caches

# This diff

This diff adds DirectoryMetadata to TreeEntry to make it easier to access the metadata for a single directory. This will allow EdenFS to query tree metadata per tree and therefore align w/ the way EdenFS requests Blob metadata.

Reviewed By: liubov-dmitrieva

Differential Revision: D58441064

fbshipit-source-id: 209317cd79eed264b79a3bb8a475238b64c6b768