Add option to display file originHash in eden glob
Summary:
This diff adds a new flag that will display the origin hash of the files returned as well as the total number of files.

The reason for this change to provide more granularity in accessing the results of globFiles via this endpoint, since previously these values were just dropped.

I am adding this at this time in order to assist in verification of the new EdenAPI offloading feature.

Reviewed By: genevievehelsel

Differential Revision: D58150230

fbshipit-source-id: 05a7ee4f6fb55cea7fee0f728772d97db4d21b2c