Implement Dtype handling for edenAPI globFiles pathway
Summary:
This diff adds support for looking up the file dtype when looking up globs via edenAPI.

The method is as follows
- look up the root tree corresponding to the revision
- for each file returned on that revision, if wantDtype is True, look up that file and get its dtype. Otherwise use DT_UNKNOWN.
- Store the above results in a temporary vector
- When all results are compiled combine the above vector into a Glob

The reason we are looking it up locally is that the server doesn't provide an api to retrieve the file dtype in an efficient manner. This approach, looking files up from the local tree, is chosen due to being more efficient than the alternative, which is making individual tree lookups to the server.

Even though we are adding some local file lookups, this is still overall a speed increase because the majority of the work is globbing all the files.

This is intended to be a semi-permanent solution. If Mononoke ever updates to support retrieving this kind of metadata in an efficient manner we would switch to that.

Changes made:
Glob creation is no longer done in SaplingBackingStore. Instead a list of GlobEntries is returned, which is further processed in EdenServiceHandler.

Reviewed By: kmancini

Differential Revision: D57915965

fbshipit-source-id: f2a701233d17eb4b9a832620ae6ec8864047e8e0