add support for the new "file_header_metadata" in Sapling Remote Api
Summary:
Add support for the new "file_header_metadata" FileAuxData's field in Sapling Remote Api

The current idea is to introduce it as a part of File Aux Data as an optional
field. None means that it is not fetched/not known. Empty metadata will just
correspond to an empty metadata blob.

In what cases are we are going to populate the new "file_header_metadata" field?

Currently it will be happening in two cases - trees received from CAS, and the field is populated in the conversion function from AugmentedTreeEntry to EdenApiTreeEntry. And in trees received from Sapling Remote Api  with  "augmented_trees":True flag.

Reviewed By: markbt

Differential Revision: D58591425

fbshipit-source-id: e6159e8982f6db014526a19a84de0f878926bf46