bookmarks_movement: Fully commit to relying on phases
Summary:
In D58245003, we started relying on phases information for determining which commits didn't need for restrictions to be checked.

To be closer to previous behaviour and pass integration tests, we didn't apply our optimization to the current head  as we [filtered it out](https://www.internalfb.com/code/fbsource/[51a77223d4a6a07aa6ccc1adf18f44497cfd916f]/fbcode/eden/mononoke/bookmarks/bookmarks_movement/src/affected_changesets.rs?lines=180).

That restricts the wins we can get from this optimization and also doesn't make a lot of sense.
If we're saying public commits don't need the extra check as it already happened, we should be able to rely on it for all cases.

In this diff, we implement a simpler way to decide which commits should have the restrictions enforced:
* We don't use the `hooks_skip_ancestor_of` config anymore as we won't need it
    * Instead, we purely rely on the boundary of commits which are public
* We don't limit the changesets being checked to `200_000` as
    * The physical restriction was due to the fact that we loaded all changesets at the same time, which we don't do anymore since D57906014
    * With this change, the only case where we would check `200_000` commits would be a push that is that large, which is unlikely and shouldn't skip restrictions

This means we will do less work, but do the work that needs to be done: checking new (draft) commits for any restriction violation.

The change is behind a jk for ease of rollback.

This change is necessary because the previous change didn't stop us from hitting the affected_changesets limit in prod for `whatsapp/android` Live Git Sync.

Reviewed By: markbt, RajivTS

Differential Revision: D58469686

fbshipit-source-id: 6737a520f7f3d2a627ce3f0ae933f3c8c1f07b04