Enable the import of namespaced refs
Summary:
When importing refs with namespace, there is a difference between the stored tag name and the bookmark name. The bookmark name exists with the entire namespace included in it, e.g., `namespaces/googlesource/refs/tags/android-platform-14.0.0_r7` while the tag name (embedded in the tag object) is just `tags/android-platform-14.0.0_r7`. This results in the namespaced tags being reported as simple tags when in fact they are annotated. This diff:
- Ensures that when we import namespaced tags, we don't add the redundant `tags/` prefix to them
- When fetching these tags as bookmarks, we report them as tags even though they do not begin with the `tags/` prefix
- When storing namespaced tags, we use the provided bookmark name as the tag name so there is no discrepancy (i.e. we store `namespaces/googlesource/refs/tags/android-platform-14.0.0_r7` in both `bookmarks` and `bonsai_tag_mapping` table)

Reviewed By: andreacampi

Differential Revision: D58668972

fbshipit-source-id: c995b0e12c3ce01b60b7c9122bfc3ee7f41dafdc