Support resolved macros in named deps in autocargo
Summary: This diff implements support for dynamically computed `named_deps` crate names in autocargo using the same mechanism that was used to support them in rust-analyzer in D58638036. Buck2 exposes a `…[named_deps]` subtarget which is a file containing the actual name of all the named deps, and when producing autocargo's JSON input from fbcode_macros we can read this and update the `named_deps` dictionary entry with the real crate name, similar to how we already update `crate` in a similar way by reading the value of `crate_dynamic` from a file.

Reviewed By: zertosh

Differential Revision: D58656554

fbshipit-source-id: 1e3de91852b1aad8d0212295b729099c6eedf984