Migrate CommitInfoTextField from VSCodeTextField to TextField
Summary:
Replace uses of VSCodeTextArea used in the comit info view for the title/summary/description editors.

This is the largest and most involved use of VSCodeTextArea, so this is a somewhat involved change.

Notably, before, we frequently had to access the inner `:part(control)` of the VSCodeTextArea shadow DOM. Several components and features have to do this, and it required passing refs which were hard to type.

By directly using a TextArea, the refs are quite easy to type, and we no longer need the "inner" textarea access.

One user of VSCodeTextArea was MinHeightTextArea, which we treat exactly like VSCodeTextArea in CommitInfoViewField, so we had to migrate that at the same time.

This diff tries to be minimal by not migrating all the tests' wrapper logic.

I see a couple of small behavior issues after this change, which I'll fix forward in later diffs.

Reviewed By: quark-zju

Differential Revision: D58594578

fbshipit-source-id: abd5aab1a4aea4175cef6560d0cbf16ae21d897c