Migrate quick commit field to TextField
Summary:
Migrate the quick commit input field to a <TextField>.

We can reduce how many type casts we need when doing this since it's much easier to type an <Input> than the <VSCodeTextField>'s shadow dom ref.

We also need to update some styles that look for `vscode-text-field` in css. In particular, the feature that hides the commit field until you hover or have focus also keeps showing the text field if you have something typed in there. We use `:not(:placeholder-visible)`, which happens to be the exact criteria of if something has been already typed.

Reviewed By: quark-zju

Differential Revision: D58608747

fbshipit-source-id: c8fdddcb153fb427abf6f38a952559d9d5ff4c1f