Fix SeeMoreContainer
Summary: Seems like we had a bug in SeeMoreContainer, which made it not really expand very much. It needs to actually limit the height. Now we actually inherit the right size. I also bump it up because it was fairly short by default and clearly people can handle the default height.

Reviewed By: quark-zju

Differential Revision: D58618939

fbshipit-source-id: 455a95285f42862eaa98b621c80edb0d152daa9a