respect the tree attributes
Summary:
respect the attributes

Returning everything with augmented manifest is small overhead since everything is stored in the same blob, but it is overhead in terms of amount of data we need to send to client.

So, let's respect the initial breakdown we have for this endpoint. The augmented digest for the tree itself is always returned.

Reviewed By: quark-zju

Differential Revision: D58677927

fbshipit-source-id: 946182f7eb99d0515e22bfcc9947d8a4756ddcff