dag: do not bump IdMap/IdDag version for virtual vertexes
Summary:
Inserting to the virtual group technically requires a VerLink version bump.

Practically it could be too annoying to handle. See D47488534 for an example.
Let's just skip bumping the verlink when inserting to the virtual group for
now. The IdDag uses indexedlog-provided storage version so it already skips
bumping.

Reviewed By: zzl0

Differential Revision: D58549205

fbshipit-source-id: 365de7f20287c56cf0f7d8eba7020d7e0335f364