Fix missing webview-ui-tookit theme variables
Summary:
Now that we no longer import from the vscode-webivew-ui-toolkit anywhere, our theme variables are no longer working. This is because of two behaviors of the library:

1. in vscode webviews, define a set of theme variables based on vscode theme variables, like `--background` as `--vscode-editor-background`.
2. in non-vscode webviews, define the same set of variables with fallback constant values corresponding to the default vscode dark theme

The mapping for (1) is here: https://github.com/microsoft/vscode-webview-ui-toolkit/blob/a1f078e963969ad3f6d5932f96874f1a41cda919/src/design-tokens.ts#L79
which we need to implement ourselves for the vscode webview

For (2), we already had a set of constants for light theme, so we just need the same list for dark theme.

Reviewed By: quark-zju

Differential Revision: D58702154

fbshipit-source-id: 5f239a450c868b9891f1d1e9a5c8eaf27bd5b691