Use <Panels> for EditStackModal
Summary:
Use the <Panels> component in the last diff for the edit stack modal. This is the only use of <VSCodePanels>, and this is the last webview-ui-toolkit component used in ISL.

I think the code is simpler after this change, since we don't need to manually selectively render panels, and xstyle works decently well for doing the custom styling.

Reviewed By: quark-zju

Differential Revision: D58699369

fbshipit-source-id: d89c645980884e90fa7eab40cc8c9ec4d8c38cd8