Introduce the RepoMetadataInfo facet
Summary:
We want to be able to run the `RepoMetadataLogger` tool in incremental mode. This will require us to have a mechanism to store the last successfully processed state. The `repo_metadata_info` table will be used for that purpose.

This diff just contains the scaffolding. The follow up diffs will populate the needed content.

AOSC diff: D58725609

Reviewed By: YousefSalama

Differential Revision: D58725241

fbshipit-source-id: cb539428805fe3e538b4cbe573fa668018fb1b24