workingcopy: disable sparse matcher for dotgit
Summary:
In the dotgit mode, the working copy is maintained by `git` which does not
understand `sl` sparse profile. This avoids an issue of
`vfs.root.join(dot_dir)` can be "bad" if `dot_dir` is `".git/sl"`.

This seems to unblock `sl status` in a submodule in a dotgit working copy.

Reviewed By: muirdm

Differential Revision: D56635187

fbshipit-source-id: ab8f80d3a978ecdc546bd503437bb7efe20cc84d