Add change_type enum for files and directories
Summary: The new version of `RepoMetadataLogger` also needs to provide information about whether a file or directory is added, deleted or modified. This diff adds an enum that will be able to provide that value. For now, since we only support the full version of metadata logger we can't have the change type information. Hence, we populate a default `Unknown` value. Once I add logic for the diff mode, the change type will get the appropriate value.

Reviewed By: andreacampi

Differential Revision: D58735656

fbshipit-source-id: 6aea8b3d3ffec0d8f1a70a5c16388dedd4aeb2df