revisionstore: add tree_aux_store
Summary:
Add `TreeAuxStore` which will be used to store aux data for trees.

Tree aux data is the augmented manifest id and size for the corresponding tree.  This can be used to fetch the tree from an augmented manifest-based alternative store.  It's expected that we will have tree aux data *before* we have the tree, as we'll use it to fetch the tree.

When fetching a tree (either from the server for from an augmented manifest-based alternative store), we will store the tree aux data for each child that is a tree in the tree aux store.  That means it will be available to fetch each of the child trees from the augmented manifest-based alternative store.

Unlike file aux data, we store tree aux data locally, rather than in a shared cache.  This prevents cache-poisoning problems with a shared cache.  We still use a rotating indexed log, as it technically a cache, and missing tree aux data can be downloaded from the server.

Reviewed By: muirdm

Differential Revision: D58561804

fbshipit-source-id: cdd9625097c7d69823bc615b460ad2fce16e94f1