easy: commit cloud: don't reuse get_workspace in references endpoints
Summary: Right now when we call `get_references` or `update_references` we first check if the workspace exists and we use `get_workspace` for that. The problem is that `get_workspace` is its own endpoint that should fail if the workspace doesn't exist but `update_references` and `get_references` shouldn't. So let's create a function aside that does this query and use it in such endpoints.

Differential Revision: D58433271

fbshipit-source-id: 27d8bdbc4e76bbfba0ff8f72ec59b2f57d852ead